Fix: change "genai_final_score" column type from int to double as planned by design

+++ src/main/java/com/netcracker/qubership/vsec/db/MyDBSheet.java
@@ -34,7 +34,7 @@ public class MyDBSheet {
                 genai_impact_score INT DEFAULT 0,
                 genai_proactivity_score INT DEFAULT 0,
                 genai_context_score INT DEFAULT 0,
-                genai_final_score INT DEFAULT 0,
+                genai_final_score DOUBLE DEFAULT 0,
                 genai_analysis_content VARCHAR,
                 genai_analysis_impact VARCHAR,
                 genai_analysis_proactivity VARCHAR,
